Move legacy-templating-system API and code paths into legacy module.
